feat: Add booking functionality
Added Booking entity, controller, service, and repository.  Updated Event, User entities and related components.  Implemented seeders for bookings.  Modified all ID type to Integer.

+++ src/main/java/com/achraf/eventhub/event/EventController.java
@@ -19,7 +19,7 @@ public class EventController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Event> getEventById(@PathVariable Long id) {
+    public ResponseEntity<Event> getEventById(@PathVariable Integer id) {
         return eventService.getEventById(id)
                 .map(ResponseEntity::ok)
                 .orElse(ResponseEntity.notFound().build());
@@ -31,7 +31,7 @@ public class EventController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Event> updateEvent(@PathVariable Long id, @RequestBody Event event) {
+    public ResponseEntity<Event> updateEvent(@PathVariable Integer id, @RequestBody Event event) {
         return eventService.getEventById(id)
                 .map(existingEvent -> {
                     event.setId(existingEvent.getId());
@@ -41,7 +41,7 @@ public class EventController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteEvent(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteEvent(@PathVariable Integer id) {
         eventService.deleteEvent(id);
         return ResponseEntity.noContent().build();
     }
